Create Nullable `IValueConverter`s, and Add Nullable Static Analyzers (#200)
* Add Nullable Static Analyzers

* Add `ICommunityToolkitIMultiValueConverter` and `ICommunityToolkitValueConverter`

* Add XML Comments

* `dotnet format`

* Add casting to Unit Tests

Takes advantage of the Nullable Static Analysis

* Update DateTimeOffsetConverter_Tests.cs

* Update StringToListConverter_Tests.cs

+++ src/CommunityToolkit.Maui.UnitTests/Converters/DoubleToIntConverter_Tests.cs
@@ -16,7 +16,7 @@ public class DoubleToIntConverter_Tests : BaseTest
 	{
 		var doubleToIntConverter = new DoubleToIntConverter();
 
-		var result = doubleToIntConverter.Convert(value, typeof(DoubleToIntConverter_Tests), ratio, CultureInfo.CurrentCulture);
+		var result = (int)doubleToIntConverter.Convert(value, typeof(DoubleToIntConverter_Tests), ratio, CultureInfo.CurrentCulture);
 
 		Assert.Equal(result, expectedResult);
 	}
@@ -27,7 +27,7 @@ public class DoubleToIntConverter_Tests : BaseTest
 	{
 		var doubleToIntConverter = new DoubleToIntConverter();
 
-		var result = doubleToIntConverter.ConvertBack(value, typeof(DoubleToIntConverter_Tests), ratio, CultureInfo.CurrentCulture);
+		var result = (double)doubleToIntConverter.ConvertBack(value, typeof(DoubleToIntConverter_Tests), ratio, CultureInfo.CurrentCulture);
 
 		Assert.Equal(result, expectedResult);
 	}
